+++ packages/form-render/src/form-items/cascader/index.tsx
@@ -1,7 +1,7 @@
 import { Cascader, type CascaderOption } from '@arco-design/web-vue'
 import { type ComponentPublicInstance, type Ref, isRef, ref } from 'vue'
 import { watchOnce } from '@vueuse/core'
-import type { DataRecord, FormItemOptions } from '../../interfaces'
+import type { DataRecord, FormItemOptions, FormItemRenderReturn } from '../../interfaces'
 
 const cache = new WeakMap()
 
@@ -12,7 +12,7 @@ function useSelectOptions(): [Ref<SelectOptions>, (value: SelectOptions) => void
   return [selectOptions, updateSelectOptions]
 }
 
-export function renderCascaderItem<T=DataRecord>(options: RenderCascaderItemOptions) {
+export function renderCascaderItem<T=DataRecord>(options: RenderCascaderItemOptions): FormItemRenderReturn<T> {
   let selectInstance: ComponentPublicInstance
   let mounted = false